+++ src/lib/lforms/lhc-form-utils.js
@@ -8,6 +8,8 @@ import * as htmlparser2 from "htmlparser2";
 import copy from "fast-copy";
 import parse from 'style-to-object';
 
+import itemControls from '../item-controls.json';
+
 const _questionnairePattern =
   new RegExp('http://hl7.org/fhir/(\\d+\.\\d+)([\.\\d]+)?/StructureDefinition/Questionnaire');
 const _sdcPattern =
@@ -1249,6 +1251,24 @@ const FormUtils = {
     parser.end();
 
     return invalidTagsAttributes;
+  },
+
+
+  /**
+   * Escapes a string for safe use as an HTML attribute.
+   * @param val the string to be escaped
+   * @return the escaped version of val
+   */
+  escapeAttribute: function (val) {
+    return val.replace(/&/g, '&amp;').replace(/"/g, '&quot;').replace(/'/g,
+      '&#39;').replace(/</g, '&lt;').replace(/>/g, '&gt;');
+  },
+
+  /**
+   * Get an object that contain a list of Item Control codes
+   */
+  getSupportedFeatures: function() {
+    return itemControls;
   }
 
 };