+++ src/lib/lhc-data.service.ts
@@ -167,21 +167,12 @@ export class LhcDataService {
 
 
   /**
-   * Check the display type of item.text
-   * @param item an item in the lforms form items array
+   * Check the display type of item.text or an answerOption.
+   * @param item an item in the lforms form items array, or an answerOption in the lforms form answers array.
    * @returns {string}
    */
-  getItemTextDisplayType(item) {
-    var format = 'plain';
-    if (item.questionXHTML && item.questionXHTML.length > 0 && this.lhcFormData.templateOptions.allowHTML) {
-      if (!item.questionHasInvalidHtmlTag) {
-        format = 'html';
-      }
-      else {
-        format = this.lhcFormData.templateOptions.displayInvalidHTML ? 'escaped' : 'plain';
-      }
-    }
-    return format;
+  getTextDisplayType(item) {
+    return this.lhcFormData?.getTextDisplayType(item) || 'plain';
   }
 
 
@@ -226,13 +217,32 @@ export class LhcDataService {
 
 
   /**
-   * Check an item's skip logic status to decide if the item should be shown
+   * Check an item's skip logic status to decide if the item is enabled
    * @param item an item
    * @returns {boolean}
    */
+  targetEnabled(item) {
+    return this.lhcFormData ? InternalUtil.targetEnabled(item): false;
+  }
+
+
+  /**
+   * Check an item's skip logic status to decide if the item is enabled and protected (not hidden from view)
+   * @param item an item
+   * @returns {boolean}
+   */
+  targetDisabledAndProtected(item) {
+    return this.lhcFormData ? InternalUtil.targetDisabledAndProtected(item) : false;
+  }
+
+
+  /**
+   * Check if the item should be displayed (enabled or disabled but protected)
+   * @param item an item
+   * @return {boolean}
+  */
   targetShown(item) {
-    return this.lhcFormData ? item._enableWhenExpVal !== false &&
-      this.lhcFormData.getSkipLogicClass(item) !== 'target-disabled' : null;
+    return this.lhcFormData ? InternalUtil.targetShown(item) : false;
   }
 
 
@@ -356,6 +366,9 @@ export class LhcDataService {
     }
     if (item._isHiddenFromView) {
       eleClass += ' lhc-hidden-from-view';
+    }
+    if (this.targetDisabledAndProtected(item)) {
+      eleClass += ' lhc-item-disabled-protected'
     }
     if (Array.isArray(item._validationErrors) && item._validationErrors.length > 0) {
       eleClass += ' lhc-invalid'