+++ test/karma/util.spec.js
@@ -900,5 +900,16 @@ describe('Util library', function() {
   })
 
 
+  describe('getSupportedFeatures', function() {
+    it('should retrieves the supported features', function(done) {
+      $.get('/lib/item-controls.json', function(ic) {
+        var target = LForms.Util.getSupportedFeatures();
+        assert.deepEqual(target, ic);
+        done();
+      }).fail(function (err) {
+        done(err);
+      });
+    });
+  });
 });